Compilable / Valid progress update
