+++ fuzzing/CMakeLists.txt
@@ -13,8 +13,8 @@ target_include_directories(fuzzer PRIVATE ${libddwaf_SOURCE_DIR}/src)
 target_include_directories(fuzzer PRIVATE ${libddwaf_SOURCE_DIR}/src/rule_processor)
 
 
-set_target_properties(fuzzer PROPERTIES COMPILE_FLAGS "-fsanitize=fuzzer,address -fprofile-instr-generate -fcoverage-mapping")
-set_target_properties(fuzzer PROPERTIES LINK_FLAGS "-fsanitize=fuzzer,address -fprofile-instr-generate -fcoverage-mapping")
+set_target_properties(fuzzer PROPERTIES COMPILE_FLAGS "-fsanitize=fuzzer,address,undefined -fprofile-instr-generate -fcoverage-mapping")
+set_target_properties(fuzzer PROPERTIES LINK_FLAGS "-fsanitize=fuzzer,address,undefined -fprofile-instr-generate -fcoverage-mapping")
 
 target_link_libraries(fuzzer
     PRIVATE ${LIBDDWAF_PRIVATE_LIBRARIES} ${LIBDDWAF_INTERFACE_LIBRARIES}
