+++ src/PWRetriever.cpp
@@ -12,476 +12,36 @@
 #include <log.hpp>
 #include <rule.hpp>
 
-void PWRetriever::PWArgsWrapper::addParameter(const ddwaf_object input)
-{
-    // Ok, let's insert them
-    const ddwaf_object* mainArray = input.array;
-    parameters.reserve(parameters.size() + (size_t) input.nbEntries);
-    for (size_t i = 0; i < input.nbEntries; ++i)
-    {
-        parameters[std::string(mainArray[i].parameterName, (size_t) mainArray[i].parameterNameLength)] = &mainArray[i];
-    }
-}
-
-const ddwaf_object* PWRetriever::PWArgsWrapper::getParameter(const std::string& paramName) const
-{
-    //TODO: cache string rendering
-    auto param = parameters.find(paramName);
-    if (param == parameters.end())
-    {
-        return nullptr;
-    }
-
-    return param->second;
-}
-
-bool PWRetriever::PWArgsWrapper::isValid() const
-{
-    return !parameters.empty();
-}
-
-PWRetriever::ArgsIterator::State::State(const ddwaf_object* args, uint32_t maxDepth) : activeItem(args), itemIndex(0)
-{
-    stack.reserve(static_cast<size_t>(maxDepth));
-}
-
-bool PWRetriever::ArgsIterator::State::isOver() const
-{
-    return activeItem == NULL && stack.empty();
-}
-
-void PWRetriever::ArgsIterator::State::pushStack(const ddwaf_object* newActive)
-{
-    if (activeItem != NULL)
-    {
-        stack.emplace_back(std::make_pair(activeItem, itemIndex));
-    }
-
-    activeItem = newActive;
-    itemIndex  = 0;
-}
-
-bool PWRetriever::ArgsIterator::State::popStack()
-{
-    if (stack.empty())
-    {
-        return false;
-    }
-
-    auto item = stack.back();
-
-    activeItem = item.first;
-    itemIndex  = item.second + 1;
-
-    stack.pop_back();
-    return true;
-}
-
-void PWRetriever::ArgsIterator::State::reset(const ddwaf_object* args)
-{
-    activeItem = args;
-    itemIndex  = 0;
-    stack.clear();
-}
-
-uint64_t PWRetriever::ArgsIterator::State::getDepth() const
-{
-    // The depth of the active parameter is the length of the stack, plus one if the active item is a container
-    const uint64_t depth = activeItem != nullptr && (activeItem->type & PWI_CONTAINER_TYPES) != 0;
-    return depth + stack.size();
-}
-
-PWRetriever::ArgsIterator::ArgsIterator(ddwaf_object* args, uint64_t maxMapDepth) : state(args, maxMapDepth)
-{
-    if (state.activeItem != nullptr && state.activeItem->type == DDWAF_OBJ_INVALID)
-    {
-        state.activeItem = nullptr;
-    }
-}
-
-void PWRetriever::ArgsIterator::gotoNext(bool skipIncrement)
-{
-    // We need to detect the first item not to loop infinitely on "hasRelevantKey"
-    for (bool isFirstItem = true; !state.isOver(); isFirstItem = false)
-    {
-        // If there is an active item we were consuming, see if there is more data
-        // Otherwise, keep following this logic
-        if (state.activeItem != nullptr)
-        {
-            if (state.activeItem->type & PWI_CONTAINER_TYPES)
-            {
-                bool wasCurrentItemMidProcessing = false;
-
-                // Startup code flow, we want to find an object to consume but not skip the first one
-                if (skipIncrement && state.itemIndex < state.activeItem->nbEntries)
-                {
-                    skipIncrement = false;
-                }
-
-                // If the child was a container, this mean we waited on an intermediary item to run on its key.
-                // We can now proceed by skipping the `hasRelevantKey` condition
-                else if (isFirstItem && state.itemIndex < state.activeItem->nbEntries && state.activeItem->array[state.itemIndex].type & PWI_CONTAINER_TYPES)
-                {
-                    wasCurrentItemMidProcessing = true;
-                }
-
-                // Do we have an additional item to consume?
-                else if (++state.itemIndex >= state.activeItem->nbEntries)
-                {
-                    state.activeItem = nullptr;
-                    continue;
-                }
-
-                bool itemAccepted = false;
-                while (true)
-                {
-                    const ddwaf_object& newItem = state.activeItem->array[state.itemIndex];
-
-                    // If yes, is it a valid item?
-                    if (newItem.type == DDWAF_OBJ_INVALID)
-                    {
-                        // If it's not, let's loop back with the top level loop
-                        break;
-                    }
-
-                    // Is there a key to run on, we may want to pause
-                    bool hasRelevantKey = newItem.parameterName != NULL && !wasCurrentItemMidProcessing;
-
-                    // Otherwise, if it's a container, we skip the item...
-                    if (!hasRelevantKey && newItem.type & PWI_CONTAINER_TYPES)
-                    {
-                        // Empty container, ignore it
-                        if (newItem.nbEntries == 0)
-                        {
-                            // If it's not, let's loop back with the top level loop
-                            break;
-                        }
-
-                        // Push the stack
-                        state.pushStack(&state.activeItem->array[state.itemIndex]);
-
-                        // Make sure that if we ignored an intermediary key, we consider the next one we may meet
-                        wasCurrentItemMidProcessing = false;
-
-                        // We want to re-run the logic and handle encapsulated containers
-                        continue;
-                    }
-                    else
-                    {
-                        // We're good!
-                        itemAccepted = true;
-                        break;
-                    }
-                }
-
-                if (itemAccepted)
-                {
-                    return;
-                }
-            }
-            else
-            {
-                state.activeItem = nullptr;
-                break;
-            }
-        }
-
-        // Was there a stack with item to consume?
-        else if (state.popStack())
-        {
-            // Poping the stack updated the active item, let's run the first part of the logic again
-            skipIncrement = true;
-            isFirstItem   = false;
-            continue;
-        }
-    }
-}
-
-void PWRetriever::ArgsIterator::reset(const ddwaf_object* args)
-{
-    if (args != nullptr && args->type != DDWAF_OBJ_INVALID)
-    {
-        state.reset(args);
-
-        // If we're a container, we need to find the first valid item
-        if (state.activeItem->type & PWI_CONTAINER_TYPES)
-        {
-            gotoNext(true);
-        }
-    }
-    else
-    {
-        state.reset(nullptr);
-    }
-}
-
-const ddwaf_object* PWRetriever::ArgsIterator::getActiveItem() const
-{
-    if (state.activeItem == nullptr)
-        return NULL;
-
-    if (state.activeItem->type & PWI_CONTAINER_TYPES)
-        return &state.activeItem->array[state.itemIndex];
-
-    return state.activeItem;
-}
-
-void PWRetriever::ArgsIterator::getKeyPath(std::vector<ddwaf_object>& keyPath) const
-{
-    keyPath.reserve(state.stack.size() + 2);
-
-    if (!state.stack.empty())
-    {
-        for (auto iter = state.stack.cbegin() + 1; iter != state.stack.cend(); ++iter)
-        {
-            ddwaf_object arg;
-            if (iter->first->parameterNameLength != 0)
-            {
-                ddwaf_object_stringl_nc(&arg, iter->first->parameterName, iter->first->parameterNameLength);
-            }
-            else
-            {
-                // Store the index
-                ddwaf_object_unsigned_force(&arg, (iter - 1)->second);
-            }
-
-            keyPath.emplace_back(arg);
-        }
-
-        if (state.activeItem != NULL)
-        {
-            ddwaf_object arg;
-            if (state.activeItem->parameterNameLength)
-            {
-                ddwaf_object_stringl_nc(&arg, state.activeItem->parameterName, state.activeItem->parameterNameLength);
-            }
-            else
-            {
-                // Store the index
-                ddwaf_object_unsigned_force(&arg, state.stack.back().second);
-            }
-
-            keyPath.emplace_back(arg);
-        }
-    }
-
-    if (state.activeItem != NULL && state.activeItem->type & PWI_CONTAINER_TYPES)
-    {
-        ddwaf_object arg;
-        if (state.activeItem->type == DDWAF_OBJ_MAP && state.itemIndex < state.activeItem->nbEntries)
-        {
-            const ddwaf_object& lastItem = state.activeItem->array[state.itemIndex];
-
-            ddwaf_object_stringl_nc(&arg, lastItem.parameterName, lastItem.parameterNameLength);
-        }
-        else
-        {
-            ddwaf_object_unsigned_force(&arg, state.itemIndex);
-        }
-
-        keyPath.emplace_back(arg);
-    }
-}
-
-bool PWRetriever::ArgsIterator::isOver() const
-{
-    return state.isOver();
-}
-
-bool PWRetriever::ArgsIterator::matchIterOnPath(const std::set<std::string>& path, bool isAllowList, size_t& blockDepth) const
-{
-    size_t stackPos          = 1;
-    const auto& currentStack = state.stack;
-
-    for (auto iter = path.cbegin(); iter != path.cend(); ++iter, ++blockDepth, ++stackPos)
-    {
-        const std::string& key = *iter;
-
-        const ddwaf_object* argValue;
-
-        // Go throught the stack ensuring we match the list
-        if (stackPos < currentStack.size())
-        {
-            argValue = currentStack[stackPos].first;
-        }
-        // Okay, the full stack match the keys, now check the active item
-        else if (stackPos == currentStack.size())
-        {
-            argValue = state.activeItem;
-        }
-        // If the active item is a container, we have one more level
-        else if (stackPos == currentStack.size() + 1 && state.activeItem->type & PWI_CONTAINER_TYPES)
-        {
-            argValue = &state.activeItem->array[state.itemIndex];
-        }
-
-        // Our path doesn't go this far, this is a problem in the case of an allowlist, not if that's a blocklist
-        else
-        {
-            return isAllowList;
-        }
-
-        //If the tree we're matching stop matching the list
-        if (argValue->parameterName != nullptr && std::string_view(argValue->parameterName) != key)
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-bool PWRetriever::Iterator::State::isOver() const
-{
-    return targetCursor == targetEnd;
-}
-
-PWRetriever::Iterator::Iterator(PWRetriever& _retriever) : 
-    retriever(_retriever), argsIterator(nullptr, retriever.max_depth) {}
-
-void PWRetriever::Iterator::reset(const std::vector<PWManifest::ARG_ID>& targets)
-{
-    state.targetCursor = targets.cbegin();
-    state.targetEnd    = targets.cend();
-
-    // If we need to run only on new targets, we find the first new target
-    if (retriever.runOnNewOnly)
-    {
-        while (state.targetCursor != state.targetEnd && !retriever.isKeyInLastBatch(*state.targetCursor))
-        {
-            state.targetCursor += 1;
-        }
-    }
-
-    if (state.targetCursor == state.targetEnd)
-        return;
-
-    argsIterator.reset(retriever.getParameter(*state.targetCursor));
-    if (!argsIterator.isOver())
-    {
-        // If we have a valid item, then we query which high tidbits we're interested in
-        updateTargetMetadata();
-    }
-    else
-    {
-        // If we don't we let the gotoNext logic handle that for us
-        gotoNext();
-    }
-}
-
-void PWRetriever::Iterator::gotoNext(bool skipIncrement)
-{
-    argsIterator.gotoNext(skipIncrement);
-    if (!argsIterator.isOver())
-        return;
-
-    // Are we done?
-    if (state.isOver())
-        return;
-
-    // If we fully consummed the last item, let's move to the next one
-    while (++state.targetCursor != state.targetEnd)
-    {
-        // If the new target isn't in the set of new targets and this isn't the first run, skip the target
-        //	We already cached the result so there is no point in running on it again
-        //	(there is an edge case when there are multiple filters on a rule but this is handled higher up)
-        if (retriever.runOnNewOnly && !retriever.isKeyInLastBatch(*state.targetCursor))
-        {
-            continue;
-        }
-
-        // Pick the new top level object for the target
-        argsIterator.reset(retriever.getParameter(*state.targetCursor));
-
-        // If the object is valid, we good. Otherwise, loop again
-        if (!argsIterator.isOver())
-        {
-            // Update the details
-            updateTargetMetadata();
-            return;
-        }
-    }
-}
-
-void PWRetriever::Iterator::updateTargetMetadata()
-{
-    const auto& fullDetails = retriever.manifest.getDetailsForTarget(*state.targetCursor);
-    currentTargetRunOnKey   = fullDetails.inline_transformer & PWT_KEYS_ONLY;
-    currentTargetRunOnValue = fullDetails.inline_transformer & PWT_VALUES_ONLY;
-}
-
-bool PWRetriever::Iterator::isOver() const
-{
-    return argsIterator.isOver() && state.isOver();
-}
-
-PWManifest::ARG_ID PWRetriever::Iterator::getActiveTarget() const
-{
-    return *state.targetCursor;
-}
-
-const std::string& PWRetriever::Iterator::getDataSource() const
-{
-    return retriever.manifest.getDetailsForTarget(*state.targetCursor).inheritFrom;
-}
-
-const std::string& PWRetriever::Iterator::getManifestKey() const
-{
-    return retriever.manifest.getTargetName(*state.targetCursor);
-}
-
-const ddwaf_object* PWRetriever::Iterator::operator*() const
-{
-    return argsIterator.getActiveItem();
-}
-
-bool PWRetriever::Iterator::shouldMatchKey() const
-{
-    if (!currentTargetRunOnKey)
-        return false;
-
-    if (argsIterator.isOver())
-        return false;
-
-    //We shouldn't match the key of the top level item
-    return argsIterator.state.getDepth() != 0;
-}
-
-bool PWRetriever::Iterator::shouldMatchValue() const
-{
-    if (!currentTargetRunOnValue)
-        return false;
-
-    return (argsIterator.getActiveItem()->type & PWI_CONTAINER_TYPES) == 0;
-}
-
-bool PWRetriever::Iterator::matchIterOnPath(const std::set<std::string>& path, bool isAllowList, size_t& blockDepth) const
-{
-    return argsIterator.matchIterOnPath(path, isAllowList, blockDepth);
-}
-
 PWRetriever::PWRetriever(const PWManifest& _manifest, const ddwaf::object_limits &limits):
-    manifest(_manifest), wrapper(),
+    manifest(_manifest),
     max_depth(limits.max_container_depth),
     internalIterator(*this) {}
 
 void PWRetriever::addParameter(const ddwaf_object input)
 {
     newestBatch.clear();
-    wrapper.addParameter(input);
 
-    if (input.nbEntries)
-    {
-        // Populate newestBatch
-        std::unordered_set<std::string> keyNames;
-        keyNames.reserve((size_t) input.nbEntries);
-        for (size_t i = 0; i < input.nbEntries; ++i)
-        {
-            keyNames.insert(std::string(input.array[i].parameterName, (size_t) input.array[i].parameterNameLength));
-        }
-
-        manifest.findImpactedArgs(keyNames, newestBatch);
+    if (input.nbEntries == 0) {
+        return;
     }
+
+    std::size_t entries = static_cast<std::size_t>(input.nbEntries);
+    const ddwaf_object* array = input.array;
+    parameters.reserve(parameters.size() + entries);
+
+    std::unordered_set<std::string> keyNames;
+    keyNames.reserve(entries);
+
+    for (std::size_t i = 0; i < entries; ++i)
+    {
+        auto length = static_cast<std::size_t>(array[i].parameterNameLength);
+        std::string key(array[i].parameterName, length);
+
+        parameters[key] = &array[i];
+        keyNames.emplace(std::move(key));
+    }
+
+    manifest.findImpactedArgs(keyNames, newestBatch);
 }
 
 bool PWRetriever::hasNewArgs() const
@@ -494,7 +54,7 @@ bool PWRetriever::isKeyInLastBatch(PWManifest::ARG_ID key) const
     return newestBatch.find(key) != newestBatch.cend();
 }
 
-PWRetriever::Iterator& PWRetriever::getIterator(const std::vector<PWManifest::ARG_ID>& targets)
+Iterator& PWRetriever::getIterator(const std::vector<PWManifest::ARG_ID>& targets)
 {
     internalIterator.reset(targets);
     return internalIterator;
@@ -503,81 +63,15 @@ PWRetriever::Iterator& PWRetriever::getIterator(const std::vector<PWManifest::AR
 const ddwaf_object* PWRetriever::getParameter(const PWManifest::ARG_ID paramID)
 {
     const auto& details = manifest.getDetailsForTarget(paramID);
-    return wrapper.getParameter(details.inheritFrom);
-}
 
-bool PWRetriever::moveIteratorForward(Iterator& iter, bool shouldIncrementFirst)
-{
-    if (shouldIncrementFirst)
-        iter.gotoNext();
-
-    while (!iter.isOver())
+    //TODO: cache string rendering
+    auto param = parameters.find(details.inheritFrom);
+    if (param == parameters.end())
     {
-        const auto& arg = manifest.getDetailsForTarget(iter.getActiveTarget());
-        if (arg.keyPaths.empty())
-            return true;
-
-        const bool isAllowList = arg.isAllowList;
-        bool matchPath         = true;
-        size_t maxDepth        = 0;
-        size_t blockDepth      = 0;
-        // The key is accepted, if:
-        //  - in the case of a blocklist, it was rejected consistently
-        //  - in the case of an allowlist, it accepted at least once
-        if ((matchPath = iter.matchIterOnPath(arg.keyPaths, isAllowList, blockDepth)))
-        {
-            if (arg.isAllowList)
-                return true;
-
-            // The blocklist matched :(
-            maxDepth = blockDepth - 1;
-        }
-
-        // Either we matched the full list and never matched (good for blocklist, bad for allowlist)
-        //		or we matched the blacklist
-
-        // Never matched the blocklist
-        if (!matchPath && !isAllowList)
-            return true;
-
-        // If we matched the blocklist, we need to first pop the stack until we get out of the restricted region
-        bool popped = false;
-
-        if (!isAllowList)
-        {
-            while (iter.argsIterator.state.getDepth() > maxDepth && iter.argsIterator.state.popStack())
-                popped = true;
-        }
-
-        iter.gotoNext(popped);
+        return nullptr;
     }
 
-    return false;
-}
-
-bool PWRetriever::runIterOnLambda(const PWRetriever::Iterator& iterator, const std::function<ruleCallback>& lambda)
-{
-    const ddwaf_object* input = *iterator;
-    //Do we have data?
-#ifndef TESTING
-    //This should be impossible with messing with the Iterator context and we can't call this function directly as it'll break the coverage
-    // I'd still like the keep the test in just to be on the safe side
-    if (input == nullptr)
-        return false;
-#endif
-
-    //We match the key of the ddwaf_object if it exists
-    if (iterator.shouldMatchKey() && lambda(input, DDWAF_OBJ_STRING, true, true))
-    {
-        return true;
-    }
-
-    if (iterator.shouldMatchValue() && lambda(input, input->type, false, true))
-    {
-        return true;
-    }
-
-    return false;
+    return param->second;
 }
 
 void PWRetriever::resetMatchSession(bool _runOnNew)
@@ -587,5 +81,5 @@ void PWRetriever::resetMatchSession(bool _runOnNew)
 
 bool PWRetriever::isValid() const
 {
-    return wrapper.isValid();
+    return !parameters.empty();
 }
